RPC API updates and integration:
* `AccountManager.listAccounts` now returns the address and not the entire StoredAccount objects.
* Added `Miner.minerStop`, `Miner.minerStart`, `Miner.mining` and `miner.Hashrate` RPC methods to the miner package for starting stopping and query miner state.

Others:

* `Response.ID` of JSONRPC 2.0 response object type changed to `uint64`.
* Removed deprecated `spell` package.
* Console executor's `PrepareContext` now collects all known RPC methods and adds them to the `ell` global JS object.

+++ rpc/jsonrpc/jsonrpc.go
@@ -24,7 +24,7 @@ type Request struct {
 	JSONRPCVersion string `json:"jsonrpc"`
 	Method         string `json:"method"`
 	Params         Params `json:"params"`
-	ID             int    `json:"id,omitempty"`
+	ID             uint64 `json:"id,omitempty"`
 }
 
 // IsNotification checks whether the request is a notification
@@ -45,7 +45,7 @@ type Response struct {
 	JSONRPCVersion string      `json:"jsonrpc"`
 	Result         interface{} `json:"result,omitempty"`
 	Err            *Err        `json:"error,omitempty"`
-	ID             int         `json:"id,omitempty"`
+	ID             uint64      `json:"id,omitempty"`
 }
 
 // IsError checks whether r is an error response
